requirements.txt
gradio>=4.0.0
openai>=1.0.0
google-generativeai>=0.3.0
python-dotenv>=1.0.0
flask>=2.3.0
flask-wtf>=1.2.1
bcrypt>=4.1.2